fix issue with read_data when accidentally placing non-tuple data into it
