+++ hhat_lang/cli.py
@@ -1,14 +1,16 @@
 """Console script for hhat_lang."""
 import sys
 import click
+from hhat_lang.evaluator import Code
 
 
 @click.command()
 def main(args=None):
     """Console script for hhat_lang."""
-    click.echo("Replace this message by putting your code into "
-               "hhat_lang.cli.main")
-    click.echo("See click documentation at https://click.palletsprojects.com/")
+    c = "main null C: (int a=(:add(1 1), :print))"
+    code_exec = Code(c)
+    code_exec.run()
+    # click.echo("See click documentation at https://click.palletsprojects.com/")
     return 0
